利用ユーザー作成用のエンドポイント追加

+++ src/entity/client_user.go
@@ -9,8 +9,8 @@ type ClientUser struct {
 	Email        string    `json:"email"`          // Eメール
 	FirebaseUID  string    `json:"firebase_uid"`   // firebase id
 	DeleteFlag   bool      `json:"delete_flag"`    // 削除フラグ
-	CreateUserID int       `json:"create_user_id"` // 作成者
-	UpdateUserID int       `json:"update_user_id"` // 更新者
+	CreateUserID int64     `json:"create_user_id"` // 作成者
+	UpdateUserID int64     `json:"update_user_id"` // 更新者
 	Created      time.Time `json:"created"`        // 作成日時
 	Updated      time.Time `json:"updated"`        // 更新日時
 }
